# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Marks",
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "https://localhost:5050"
    },
    {
      "url": "https://localhost:5051"
    }
  ],
...
  "paths": {
...
    "/users/get_login": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get user`s login",
        "description": "Get user`s login from JWT access token payload",
        "operationId": "getlogin",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "login": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Wrong token"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/users/get_user": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get user`s info",
        "description": "Get user`s login from JWT access token payload",
        "operationId": "getuser",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "login": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Wrong token"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/users/refresh": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Refresh JWT access token using refresh token",
        "description": "",
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "402": {
            "description": "Wrong token"
          },
          "403": {
            "description": "Wrong refresh token"
          }
        },
        "security": [
          {
            "cookieRefresh": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access-token"
      },
      "cookieRefresh": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh-token"
      },
      "marks_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://localhost:5050/users/login",
            "scopes": {
              "write:marks": "modify marks in your account",
              "read:marks": "read your marks",
              "write:students": "modify students in your account",
              "read:students": "read your students"
            }
          }
        }
      }
    }
  }
}